pyproject.toml
[project]
name = "chatrfp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "gunicorn>=23.0.0",
    "langchain-community>=0.3.20",
    "langfuse>=2.60.2",
    "openai>=1.68.2",
    "orjson>=3.10.15",
    "pandas>=2.0.3",
    "pypdf>=5.4.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "qdrant-client>=1.12.1",
    "streamlit>=1.40.1",
    "streamlit-pdf-viewer>=0.0.21",
    "uvicorn>=0.33.0",
]

[dependency-groups]
local = [
    "fastembed>=0.4.2"
]
prod = [
    "fastembed-gpu>=0.4.2",
]
